Separate JobClient into SyncJobClient and AsyncJobClient decorator Cleanup the job client code dealing with closing Request and Response objects by using Closeable and try-with-resources construct Fix and update the unit test
